[REFACTOR] common 폴더 개설 및 파일 이동
